remove memory snapshot feature
